workflows/react_monolith/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3>=1.40.0
httpx
langchain
langgraph
langchain-openai
traceloop-sdk
python-dotenv
langgraph-checkpoint-aws